MB-51556: Remove Monotonic<> labels in preparation for new style labels
Remove the current std::string labels we assign to Monotonic<> or
AtomicMonotonic<>, so that we can update the way we create labels for
Monotonic<> template so that they are more efficient with respects to
memory usage.

Follow up patches will re-introduce the labels using the newer more
efficient method, after montonic.h in platform repo us been updated.

Change-Id: I82b17b486bdc6604b47a8ea7842b81f0d1605251
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/173361
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>