MB-43027: Always enqueue vBucketState when adding failover entry
A bug existed in a previous release in which a failover table entry
was created but the vBucket state was not enqueued into the
CheckpointManager. It was then possible for a write occur which
went to disk without updating the failover table in the vBucket state.
This caused a potential dataloss scenario. Whilst that bug has been
fixed, it would be ideal if we had a tighter coupling between failover
table entry creation and vBucket state enqueueing to prevent similar
bugs from occuring in the future.

Add a VBucket function to create a new failover table entry which also
enqueues a vBucket state into the CheckpointManager.

One note-worthy impact of this change is that we may enqueue multiple
subsequent set_vbucket_state Items into the Checkpoint. This is already
a scenario that we must deal with and the highest cas set_vbucket_state
should be the one that the flusher persists (as they share the same
seqno).

Change-Id: I16357d837776ba33942fe22875c8195a7c7deb10
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/173429
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>