MB-51556: Re-add Monotonic label to queuedSeqnoAcks
Re-add label for instances Monotonic<> stored in queuedSeqnoAcks, using
the new interface for generating a label for each monotonic.

In SeqnoAckQueueLaballer, only store the node name string if we have
CB_DEVELOPMENT_ASSERTS enabled. As having a std::string per Monotonic<>
is an expensive overhead (at least 32 Bytes for an empty std:string and
a node name is unlikely to fit within the SSO in most instances).

Change-Id: I54eb1fa1c6745bafc1e729327f36354eca0f37ef
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/173469
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>