MB-51805: Set PassiveStream::cur_snapshot_prepare on new stream
We use the cur_snapshot_prepare member to determine if we should
notify the PDM of any Memory snapshots. It is set when we see a
prepare in any snapshot. Consider the following snapshot:

[1:Prepare(A), 2:Mutation(B)] Type = Memory

And the following sequence of events:

1) SnapshotMarker (1-2) Type = Memory
2) Prepare (1)
3) Prepare (1) persisted
4) Node restarts
5) Stream reconnects
6) SnapshotMarker (1-2) Type = Memory
7) Mutation (2)

In normal operation the PDM is notified when we received the item at the
snapshot end (i.e. Mutation (2) at step 7). In this case, as the node
has been restarted, the information that the snapshot contains a prepare
is lost (cur_snapshot_prepare is not set until the stream sees a
prepare.

To solve this, we can fix the cur_snapshot_prepare state on
PassiveStream acceptance. The PDM already avoids acking back the same
seqno, so notifying an extra snapshot shouldn't matter, and even if we
did ack back the same seqno, the ADM should already deal with weakly
monotonic acks as we ack back the HPS on stream connection.

Change-Id: Ibc298db0cfad0c3d61f3fa7aeb72a79f1c092be3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/173913
Well-Formed: Restriction Checker
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>