MB-51805: Add persistedPrepareSeqno(maxPrepareSeqno) to makeVBucket()
We'll need to pass this into the CheckpointManager to correct the
state post-warmup in a future change. It might be the case that
we shutdown mid-snapshot and fail to update the highPreparedSeqno
on disk when we complete the snapshot if the rest of the snapshot
does not contain a prepare. Should we then restart, the HPS could
go backwards which could cause the loss of durable writes.

The name maxPrepareSeqno was chosen here as persistedPrepareSeqno
does not convey that the seqno being used is the seqno of the
highest prepare stored on disk (regardless of whether or not we
have acked it).

Change-Id: I436f5b9d1b349febe62deb2569310175eb1a15b3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/173914
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>