MB-51860: Don't try to create log directory in constructor
Don't try to create the audit log directory in the audit
configuration as we'll try to create in the audit write
thread as part of updating the configuration. An "invalid"
configuration is not fatal, it just means that we'll drop
all audit events and increment the audit_dropped counter
that ns_server monitors.

Change-Id: Ic2b09ca603dfb029670a916c972796828817f7db
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/173805
Well-Formed: Restriction Checker
Tested-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>