MB-51880: Use sized_buffer::data() instead of sized_buffer::buf (XAttr)
In preparation for switching from cb::sized_buffer to std::span,
remove direct uses of the 'buf' and 'size' member variables, using
data() and size() instead.

In the case of xattr::Blob::get(), this change highlights an error in
the const-ness of that method - the method is marked const, but it
returns a writable view into an XATTR value (cb::char_buffer is a
non-const view).

Fix that by marking get() as non-const, which in turn requires
tweaking usage elsewhere.

Change-Id: I9de01ffa30cb7043a4ef0fcc4e053537a9a8399a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/173904
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>