MB-51805, MB-51806: Inherit HPS from previous Checkpoint
MB-51805:

Whilst the next patch in this chain fixes the issue on the neo
branch, a cherry-pick of this set of changes do not solve this
issue on the master branch.
https://review.couchbase.org/c/kv_engine/+/173044 removed the
Checkpoint re-use optimization on the master branch on which the
previous change (to load the PPS from disk into the CheckpointManager)
relies. On the master branch, when we receive our first SnapshotMarker
we create a new snapshot in the CheckpointManager which will actually
create a new Checkpoint. The HPS value that we have loaded from disk
and passed into the Checkpoint will no longer be correct.
Whilst we would normally just fix this on the master branch, we are
making the same change for MB-51806 anyway and I feel that it's
important to highlight the dual purpose of the change here.

MB-51806:

Currently we only update the highPreparedSeqno on disk if we are
persisting the item at the end of a snapshot. This is correct, but
fails to take into account that a Checkpoint only sets it's HPS
value when it sees a prepare. We could have a flush batch such that
we persist no item at a snapshot end for Checkpoints containing
prepares and the HPS is not moved when it should be. E.g.

[1:Pre(key1), 2:Mutation(key2)][3:Mutation(key2)]
             ^
         HPS = 1

When visited by the flusher the HPS of 1 is not persisted. This allows
for a scenario where a replica node restarts and the HPS goes backwards
as this node would still ack 1 back to the Producer. If the cluster
has two replicas and the other replica is behind, it's possible for it
to be picked for promotion over this node. See MB for full scenario.

This change fixes this issue as the second Checkpoint inherits the
highPreparedSeqno value from the previous one, and rather than finding
no value to flush, it finds the correct value of 1.

Change-Id: I06f43c18975049187705efcb967f2d9c1fa5c782
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/173918
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>