MB-51877: Create a sloppy compute unit gauge
The first implementation of the throttling of user commands
should just use cap the number of CUs used on a per sec basis.

For simplicity for the initial implementation we'll be using
a sloppy gauge which gets moved by the callback which drives
the "current clock" im memcached (1 sec callback).

It works by having 60 "slots" keeping track of the CU used
within the slot. All updates and checks happens in
the "current" slot, and at a given interval "someone" (the
callback from the clock thread) reset the next slot and move
the current slot forward.

Change-Id: I2a7b99f311846196f7fcef892016d57554ccbda5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/174026
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>