MB-51925: Don't allocate for access logs unless necessary during Warmup
Currently we allocate memory for access logs during warmup before we
use them. We may not even use them if we can enable traffic immediately.
Move the construction of them to the point at which we need them.

We also don't need the access logs after we've loaded the data in them
so we should free up the memory after we're done with them.

Change-Id: Ib271eac227ad1b055484758781bbaa633ae1a625
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/174054
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>