MB-35297: RangeScan Create and Cancel using mcbp
Add validation and executor code so that the RangeScan
create and cancel can be invoked using the memcached binary
protocol (opcodes 0xDA and 0xDC).

The create command requires the user specifies a JSON value
that encodes the numerous parameters of a scan. Validation
ensures that the value is JSON and that the collection-ID is
setup ready for privilege checks.

Cancel is much simpler, with only one parameter available.

The added markdown provides full details of each new command.

Note: RangeScan continue (0xDB) is a separate patch.

Change-Id: Ib1a056f6990f982b101d33ec3b03b7a13918e96b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/175057
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>