Add support for requesting information for a single bucket
Previously one would need to request the detailed bucket stats
for all buckets and then look up the bucket on the client side.
This adds extra work for all of the unit tests which needs to
check for individual bucket state information.

This patch adds the ability to request information for a
single bucket.

Change-Id: Ib6ec96fc67092c71cbd78421f904cd62573969e4
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/174291
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>