MB-51877: Let SCUGT::tick rollover overuse
The way we're planning to use the Gauge is to just check if
there is any room available before starting to execute an
operation, and once the operation completes we'll account for
the data being used.

A problem with the previous implementation was that you could
overuse your quota a lot, and every time you entered a new time
slot you would be free to do it over again.

If you for instance set up a limit of 5k CU/s the max document
size you can operate on is 5M. There was nothing stopping you
from running a ton of parallel operations on 20MB documents
(as in the beginning of the time slot there was 5K CU available
so you could start all operations) and when all of they complete
you have used your quota for a long time, but instead you would
start with clean sheets at the next second boundary.

This patch let the over-use of the quota roll over to the next
timeslot.

Change-Id: I43a9a2a3dbcb61fa4ce9aaaea3a09894ec85f356
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/174293
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>