MB-51914: Set a max limit on #connections to a bucket
Only allow external clients to select a bucket if the total
number of connections bound to the limit is lower than 600.

Once the requirements for the limit (is it a global limit
or may it be set differently across the buckets etc) we
need to implement a way to set/tune the limit.

Change-Id: Ife4a37bd2e34f7a8e0574145730797a068539ece
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/174391
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>