MB-35297: Validate vbucket has a matching vb-uuid
A vbucket could change uuid in the window between checking
the in-memory vbucket and the RangeScanCreateTask opening the
vbucket.

The RangeScanCreateTask must check the vb-uuid from the snapshot
matches when snapshot-requirements are defined.

A vbucket could change uuid whilst a scan is idle or waiting to
run in a response to continue.

The RangeScan::continueScan path must check the vbucket is still
as expected, this is done along side the existing "active" check for
each key callback.

Change-Id: Ia8a28113f810662e566dc639b4dab9e70ed9d1b8
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/174993
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>