MB-51914: Remove locks for throttle-list
Using locks for the throttledConnections member in the bucket isn't
really needed as a connection (and all its cookies) gets pinned to
a given front end thread during connection. Once bound to a connection
it will NEVER switch to a different connection.

Given that the list of throttled connections may be used in a lock
free manner as long as we only operate on the list from a given
thread; which is what we do: the front end thread.

Change-Id: I72ce2b33550069ec7becb73a71d6b8145d5d646d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/174515
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>