MB-50984: Remove test boilerplate introduced in MB-36146
As per offline discussion with the author, the related unit test is
invalid as it relies on conditions that are impossible to hit in
production.

In particular, the following snippet is part of the production code:

bool CheckpointManager::removeCursor(const std::lock_guard<std::mutex>& lh,
                                     CheckpointCursor& cursor) {
..
    /**
     * The code bellow is for unit test purposes only and is designed to inject
     * code to simulate a race condition with the destruction of a cursor. See
     * for more information MB-36146
     */
    if (runGetItemsHook) {
        queueLock.unlock();
        runGetItemsHook(cursor, vb.getId());
        queueLock.lock();
    }

    return true;
}

The above function is always executed under CM::queueLock. The lock is
temporarily released only in the unit test.

Change-Id: I1d7cb7c66831ff5c4e0ecff9607aa49665cabe03
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/174626
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>