MB-52067: Handle not found cas op status during counter store
Counter operations have a feature that that can create a counter
document with the provided initial value if the document does not
already exist. The arithmetic state machine currently deals with
this by first looking up if a document exists or not, and then
issuing either a cas (for update atomicity) set if it already
exists or an add if the document does not exist. There exists a
race condition in this logic though as follows:

1) GetItem runs for client A and finds the document
2) Document is deleted by client B
3) AllocateNewItem runs for client A (this copies the cas of the
   existing item to perform an atomic operation)
4) StoreItem runs for client A and finds that the document has
   been deleted

At step 4 when we find that the document is in a deleted state in
the HashTable we return up the stack a not found/no such key status
as this is a cas operation. The arithmetic state machine does not
attempt to retry the counter increment though, it instead passes
the not found status up the stack and back to the caller. This goes
against the create if non-existent feature.

Handle a not found status in the same way that we handle a cas
mismatch in the arithmetic counter state machine.

Change-Id: I525f133573da83f371cab33cc8f29efc0158ce04
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/174735
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Restriction Checker