MB-51606: Fixup ADM state on transition to PDM
We must fixup the ADM state before creating a PDM because the PDM
wants to only look at the trackedWrites of the ADM, IIRC because
of some lock order inversion issue. There exists an issue with the
recent change (https://review.couchbase.org/c/kv_engine/+/174464)
to allow notification of persistence to dead vBuckets though.

Recall that an active becoming a replica does so via the dead
state. If we notify persistence and that causes a write that has
already been acked to be moved from trackedWrites to the resolvedQueue
then we do not currently move that write back to trackedWrites for
the PDM construction if it happens while the vBucket is in the dead
state.

Add the required fixup call to the point at which we decide that we
have an ADM and need a PDM.

Change-Id: I196dcb0e61dbc83a225566207807e2105c495ed3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/174799
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>