[BP] MB-39344: Fix computation in Checkpoint::getHighSeqno()
The user of this API (CM::registerCursor() only at the time of writing)
expects that the function returns high-seqno meaning the seqno of the
last non-meta item in the checkpoint.
With regard to that the function is broken as it might return
(high-seqno + 1) in the case where the last items in the checkpoint is
a is a set-vbstate.

I've tried to break the CM::registerCursor() behaviour with the
original C:getHighSeqno() code on local tests with no success.
Reason is that the logic at CM::registerCursor() is resilient to the
misbehaviour that we correct here. See test for details.

Change-Id: Ia77304dfc1a253407e044740cfcecea29b41a3e2
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/194823
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>