[BP] MB-39344: Add test coverage for Checkpoint::getMinimumCursorSeqno
The function's correctness it's crucial to our cursor-registering
logic. The patch improves comments to clarify the function's behaviour.
Also I add full test coverage on how the various item-removal paths in
checkpoint affect the result of the function and the consequent
behaviour of cursor-registering. That is all in preparation of changes
in the ItemExpel area.

A couple of collateral issues have arised from tests, see test code for
details. Issues are being fixed in follow-up patches.

Change-Id: Ib0a759851ea7eced250798af9f41c857fa076819
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/194825
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>