MB-52065: Add notify_io_complete trace span
Add a trace span for when notify_io_complete is being
called. This call is intended to be fast, but we don't
know _when_ it is being called.

We've seen occurrences in the linked bug where the
request span covers a long period of time, but the
individual execute spans are really short. Given
that we don't have any other spans recorded for the
operation we have no idea _why_ the operation is
so slow.

Change-Id: I7f5965678abe8b746d9a34e51baeb398e7132587
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/174981
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>