[BP] MB-39344: Allow expelling all non-meta items in checkpoint
- Notes -

. As in the master branch, this BP triggers MB-53055. I include the
  fix in this BP,   https://review.couchbase.org/c/kv_engine/+/179284.
  Small production code change, covered by unit tests. I include the
  original commit message below.

. As in the master branch, the original patch triggers MB-53570. I
  include the fix in this BP,
  https://review.couchbase.org/c/kv_engine/+/179618. Small production
  code change, covered by unit tests. I include the original commit
  message below.

. Differently from Neo, in master the original change was submitted
  into a code base where there is no checkpoint creation by "max num
  items". This backport interferes with that feature in some
  ep_testuite tests. They are all benign test issues that I fix here by
  just disabling Expel on those particular tests.

- Original commit message of the main BP -

The historical problem of unrecoverable OOM on replica checkpoints has
been highly relieved in Neo compared to the previous releases, but not
totally solved.

In particular, low-end envs with low memory quotas might experience
hard OOM in scenarios like simple data loads or rebalance-in. That
happens in cases where the doc size is high enough to eat the CMQuota
by a few docs spread across multiple vbucekts, eg:
 1. BucketQuota 100MB, CMQuota 30MB
 2. Doc size 10MB
 3. 4 vbuckets
 4. Load 10 docs on Node1 - no replica yet
 5. RebIn Node2

At (5) Node2 might receive ~ 3 docs across multiple replica vbuckets,
1 doc per vbucket. That allocates ~ 30MB in checkpoints (ie CMQuota
full), but:
 a. Replica checkpoints are open and won't close unless the node
    receives some new snapshots over those streams immediately. So
    checkpoints can't be released by CheckpointRemoval.
 b. We can't Expel the items in those checkpoints as we can't remove
    the last item in a checkpoint

We fix the problem by solving (b). With this patch ItemExpel is
capable to remove all the items in a checkpoint if memory recovery
requires so. With all items we refer to non-meta items only, so the
integrity of the checkpoint is preserved.

- Original commit message of the MB-53055 fix -

MB-53055: Fix Checkpoint::isEmptyByExpel() semantic

As part of MB-39344 we have fixed the semantic of
Checkpoint::getHighSeqno() from (erroneously) accounting for seqnos of
also meta-items to accounting only seqnos of non-meta items.

But we missed to fix Checkpoint::isEmptyByExpel() to. In the old
definition of Checkpoint::highSeqno a checkpoint is empty only if it
shapes like [e cs). In the new definition of C::highSeqno also a
checkpoint like [e cs vbs) is empty.

That inconsistency is failing QE tests at CM::registerCursor() as we
essentially end up executing non-empty checkpoint seqno-checks on empty
checkpoints.

- Original commit message of the MB-53570 fix -

MB-53570: Make CM::registerCursor resilient to first empty checkpoint

As part of MB-39344 we have added some sanity checks at
register-cursor. As seen in MB-53570 it turns out that some of those
checks are too strict and miss to consider that a CheckpointList like

  [disk | e:1 cs:1 x vbs:3) [memory | e:3 cs:3)

is legal on a vbucket after some state transition. See new unit test
for details.

The failure here is failing many QE tests, so it's a blocker.
This is a quick fix where we just relax the sanity checks. The existing
logic in CM::registerCursor handles the case with no further change,
proved by a specific unit test.

Some suboptimal behaviour has been spotted with regard to potential
unnesessary backfills triggered though. That will be addressed in
MB-53616.

Change-Id: I21f0efdc94c709575a6f1a105f2e58408f5e46f3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/194829
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>