MB-39344: Allow expelling all non-meta items in checkpoint
The historical problem of unrecoverable OOM on replica checkpoints has
been highly relieved in Neo compared to the previous releases, but not
totally solved.

In particular, low-end envs with low memory quotas might experience
hard OOM in scenarios like simple data loads or rebalance-in. That
happens in cases where the doc size is high enough to eat the CMQuota
by a few docs spread across multiple vbucekts, eg:
 1. BucketQuota 100MB, CMQuota 30MB
 2. Doc size 10MB
 3. 4 vbuckets
 4. Load 10 docs on Node1 - no replica yet
 5. RebIn Node2

At (5) Node2 might receive ~ 3 docs across multiple replica vbuckets,
1 doc per vbucket. That allocates ~ 30MB in checkpoints (ie CMQuota
full), but:
 a. Replica checkpoints are open and won't close unless the node
    receives some new snapshots over those streams immediately. So
    checkpoints can't be released by CheckpointRemoval.
 b. We can't Expel the items in those checkpoints as we can't remove
    the last item in a checkpoint

We fix the problem by solving (b). With this patch ItemExpel is
capable to remove all the items in a checkpoint if memory recovery
requires so. With all items we refer to non-meta items only, so the
integrity of the checkpoint is preserved.

Change-Id: I21f0efdc94c709575a6f1a105f2e58408f5e46f3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/175051
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>