MB-35297: Fix shutdown crash whilst a scan was idle
During shutdown the RangeScan objects will get destructed as
part of the ~VBucket path (which comes from KVShard's 'vbuckets'
vector destructing).

The RangeScan will free the ScanContext triggering a close of the
underlying KVStore, which prior to this patch had been destructed
already due to the ordering of KVShard's members.

Change-Id: I8fcee2a405f08b1017b555c9c5c0dd938745e5df
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/175211
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>