MB-35297: Cap the number of RangeScanContinueTasks
Prior to this commit RangeScans create one new task for each
continue or cancel request, allowing for RangeScans to utilise
all available AUXIO threads.

This commit caps the number of tasks that can exist by tracking
how many are in existence. A task can only be created if there
is capacity. The RangeScanContinueTask itself is adjusted so that
it keeps rescheduling whilst there are RangeScan objects that need
processing.

By default this commit will ensure that one AUXIO thread would be
free by only ever creating one less than the number of AUXIO threads
that exist.

Change-Id: I6e8e4d4bf9b6c608b233876749b9d7fb58b27908
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/175330
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>