CBD-6348: [BP] Changes required for folly v2022.02.14.00
This is a cherry-pick from trinity, but the ChangeID is changed
to allow linking with the tlm change

Folly v2022.02.14.00 has removed the method getUniquePtr(), replaced
with as_lock().

Unfortunately our previous version of folly (v2020.09.07.00) did not
yet have the as_lock() function, so we cannot update kv_engine in
advance of the folly update - we must change the method in lock-step
with the upgrade.

Change-Id: If7fd9aea96f4c23092009a985d0ad0d839aac596
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/234760
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Trond Norbye <trond.norbye@couchbase.com>