MB-51916: Add support for refuse data ingress from clients
Add support for the data plane to tell a bucket that it should
not accept data ingress from a client into a bucket. Once
enabled a client may only retrieve (and delete) documents.

Note that this does _NOT_ affect DCP as we don't want to
end up in a situation where we can't store the requested
amount of replicas.

Change-Id: Ib4de82fde4cea46292c8fc01d56ca9ebe952aff1
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/175151
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>