MB-52158: Check for privilege in RangeScan continue/cancel
In this patch Privilege::Read is checked, but will change to
Privilege::RangeScan once available.

A range-scan requires that the user has the RangeScan privilege
somewhere in the "path" to the collection. For RangeScan create
this is checked by mcbp_privileges as the incoming command states
which collection the scan is against.

For continue/cancel the engine must do the checks as it needs to
locate the scan and then the scan's collection.

A "bonus" of this is that the collection is also existence checked
during the checkPrivilege method, allowing continue to 'force'
cancel the scan if the collection rather than waiting for the scan
to move forward (load a key) and fail in range_scan_callbacks.cc

Change-Id: I3e1b6ad043e9d377aa897b8a2a64a637f36ac7a9
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/175752
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>