MB-51876: Unit tests to verify metering [1/n]
The quiet ops (and the getk variants) aren't being used
by our SDKs and we don't want them to be introduced in a
serverless deployment. Add unit tests to verify that they
return NotSupported.

Start adding unit tests (and update the implementation)
for each of the available commands to be sent to the
server and verify that we meter them.

Change-Id: Id5002ae3b7577501daec42a96f45e5c87bfc0594
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/175248
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>