Revert initial fix for MB-51606
Since recent changes to notify persistence to ADM when vBucket is dead
(MB-51606), we attempt to update the ADM to account for the new
persisted seqno. This can cause a crash if there are ActiveSyncWrite
objects in trackedWrites which do not have any chain information - see
MB-52169.

As per discussions with QE, the latest bug (MB-52169) has a larger
impact than the original MB-51606, and hence in the short-term (while
we investigate the best way to fix both issues) we will revert the
original fix for MB-51606.

This commit reverts the following commits:

* Revert "MB-51606: Fixup ADM state on transition to PDM"
  This reverts commit 3a16c9c3800ef883db73727716f909f513412821.

* Revert "MB-51606: Don't throw if PDM of dead vBucket tries to ack"
  This reverts commit 793e61741e893a87d7adaebb95857cdb4cbd2efc.

* Revert "MB-51606: Update seqno ack to send outside state lock"
  This reverts commit 6912b02bb8aad06eb439a071668a076a1a205ba4.

* Revert "MB-51606: Pass VB State lock to VBucket::sendSeqnoAck()"
  This reverts commit 8926882d82a402ee77aac869ebda4676b93ba280.

* Revert "MB-51606: Pass VB State lock to PDM::notifySnapshotEndReceived()"
  This reverts commit 0f32a5d174a4aecfacd8c057d33d0a6818b03715.

* Revert "MB-51606: Pass VB State lock to DM::notifyLocalPersistence()"
  This reverts commit 90245a1f6ddfb434aaefb6a6f871f3bfe4fca78b.

* Revert "MB-51606: Notify DM for persistence of dead vBuckets"
  This reverts commit 1135ed3fa589ab1024bf09e7a7bca7ac6f6106f8.

* Revert "MB-51606: Split setupConsumerAndPassiveStream fn into components"
  This reverts commit 4552aab557d0d18d9bce26d220b73723bf432c0f.

Change-Id: Ie0336e64edcfa2e1d2f9976e6f9187190d73b1d8
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/175268
Well-Formed: Restriction Checker
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>