MB-35297: Successful RangeScan changes state to Completed
Initial motivator for this change was to improve the INFO logging
of a RangeScan, capturing when one is created and when it finally
goes away, with the desire to log the state.

Prior to this commit, all scans end up "Cancelled", so ~RangeScan could
only log that state.

An extra state (Completed) is added enabling the differentiation of
scans that were cancelled, enabling ~RangeScan to log the correct
state.

Some clean-up of the state transitions included as a lot of reviewing
of the state code occurred in this change. Resulting in removing the
needless switch from Continuing->Idle for a successful scan. Now that
case changes from Continuing->Completed.

Change-Id: I21629765f48e9ff7cd69697e095a1155c104ff51
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/175753
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>